Add service and minscore options
The service option allows specifying a preferred service instead
of random.

The minscore option allows specifying the minimum score for the
preferred service (applies to the random service if a preferred
service is not specified).

A minscore of 9999 overrides the enlistment roll.

A minscore of 8888 overrides the draft with the preferred service
(the character is still treated as having been drafted, but the
preferred service is chosen).

These special values allow generating characters that are a specific
service.

Signed-off-by: Frank S. Filz <ffilzlnx@mindspring.com>