Add minterms, maxterms, and maxchars options
With the minterms option, the generator will not accept a character
that musters out with less than the specified number of terms. It also
assures the character doesn't voluntarily muster out before that time.
The default is 1, and if left at default, will still allow deceased
characters.

With the maxterms option, the generator will limit the maximum number
of terms for a character to the specified amount. Mandatory
re-enlistment will not be checked for characters that hit the limit
(this allows a GM to specify the maximum age of PCs or NPCs).

The maxchars option will limit the maximum number of attempts to
find a requested character. The default is 10000 and bounds the time
spent attempting to generate a requested character.

Signed-off-by: Frank S. Filz <ffilzlnx@mindspring.com>